@@ -33,9 +33,9 @@
     "example": "examples"
   },
   "engines": {
-    "node": ">=6.11.1",
+    "node": ">=8.9.0",
     "npm": ">=5.3.0",
-    "yarn": ">=0.27.5"
+    "yarn": ">=1.5.1"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -52,7 +52,7 @@
     "eslint-plugin-flowtype": "^2.35.0",
     "eslint-plugin-react": "^7.2.0",
     "eslint-plugin-react-native": "^3.0.1",
-    "flow-bin": "^0.52.0",
+    "flow-bin": "^0.85.0",
     "prop-types": "^15.5.10",
     "react": "^15.4.2",
     "react-native": "^0.41.2"
