@@ -1,13 +1,9 @@
 /**
- * Sample React Native App
- * https://github.com/facebook/react-native
  * @flow
  */
 
- // #region imports
-import React, {
-  Component
-}                         from 'react';
+// #region imports
+import React, { Component } from 'react';
 import {
   AppRegistry,
   StyleSheet,
@@ -16,18 +12,17 @@ import {
   View,
   TouchableHighlight,
   ToastAndroid,
-  Image
-}                         from 'react-native';
-import Beacons            from 'react-native-beacons-manager';
-import {
-  Avatar
-}                         from 'react-native-elements';
-import moment             from 'moment';
-import beaconIMAGE        from './images/beacons/ibeacon.png';
-import altBeaconIMAGE     from './images/beacons/altbeacon.png';
-import eddystoneURLIMAGE  from './images/beacons/eddystoneURL.png';
-import eddystoneTLMIMAGE  from './images/beacons/eddystone_TLM.png';
-import eddystoneUIDIMAGE  from './images/beacons/eddystone_UID.png';
+  Image,
+} from 'react-native';
+import Beacons from 'react-native-beacons-manager';
+import { Avatar } from 'react-native-elements';
+import moment from 'moment';
+import beaconIMAGE from './images/beacons/ibeacon.png';
+import { hashCode } from './helpers';
+// import altBeaconIMAGE from './images/beacons/altbeacon.png';
+// import eddystoneURLIMAGE from './images/beacons/eddystoneURL.png';
+// import eddystoneTLMIMAGE from './images/beacons/eddystone_TLM.png';
+// import eddystoneUIDIMAGE from './images/beacons/eddystone_UID.png';
 // #endregion
 
 // #region flow types
@@ -38,14 +33,14 @@ type DetectedBeacon = {
   minor?: number,
   proximity?: string,
   rssi?: string,
-  distance?: number
+  distance?: number,
 };
 
 type Section = {
   key: number,
   data: Array<DetectedBeacon>,
   title: string,
-  sectionId: string
+  sectionId: string,
 };
 
 type Props = any;
@@ -55,20 +50,21 @@ type State = {
   uuid?: string,
   identifier: string,
   // all detected beacons:
-  beacons: Array<Section>
+  beacons: Array<Section>,
 };
 // #endregion
 
 // #region constants
 //  const UUID         = '7b44b47b-52a1-5381-90c2-f09b6838c5d4';
-const IDENTIFIER   = '123456';
-const TIME_FORMAT  = 'MM/DD/YYYY HH:mm:ss';
+const IDENTIFIER = '123456';
+const TIME_FORMAT = 'MM/DD/YYYY HH:mm:ss';
+const UUID = '7b44b47b-52a1-5381-90c2-f09b6838c5d4';
 
-const RANGING_TITLE               = 'ranging beacons in the area:';
-const RANGING_SECTION_ID          = 1;
-const MONITORING_ENTER_TITLE      = 'monitoring enter information:';
+const RANGING_TITLE = 'ranging beacons in the area:';
+const RANGING_SECTION_ID = 1;
+const MONITORING_ENTER_TITLE = 'monitoring enter information:';
 const MONITORING_ENTER_SECTION_ID = 2;
-const MONITORING_LEAVE_TITLE      = 'monitoring exit information:';
+const MONITORING_LEAVE_TITLE = 'monitoring exit information:';
 const MONITORING_LEAVE_SECTION_ID = 3;
 // #endregion
 
@@ -84,17 +80,28 @@ class BeaconsDemo extends Component<Props, State> {
 
   state = {
     // region information
-    uuid: null,// UUID,
+    uuid: UUID,
     identifier: IDENTIFIER,
     // all detected beacons:
     beacons: [
-      {key: 1, data: [], title: RANGING_TITLE,          sectionId: RANGING_SECTION_ID},
-      {key: 2, data: [], title: MONITORING_ENTER_TITLE, sectionId: MONITORING_ENTER_SECTION_ID},
-      {key: 3, data: [], title: MONITORING_LEAVE_TITLE, sectionId: MONITORING_LEAVE_SECTION_ID}
-    ]
+      { key: 1, data: [], title: RANGING_TITLE, sectionId: RANGING_SECTION_ID },
+      {
+        key: 2,
+        data: [],
+        title: MONITORING_ENTER_TITLE,
+        sectionId: MONITORING_ENTER_SECTION_ID,
+      },
+      {
+        key: 3,
+        data: [],
+        title: MONITORING_LEAVE_TITLE,
+        sectionId: MONITORING_LEAVE_SECTION_ID,
+      },
+    ],
   };
 
-  componentWillMount() {
+  // #region lifecycle methods
+  componentDidMount() {
     //
     // ONLY non component state aware here in componentWillMount
     //
@@ -108,15 +115,14 @@ class BeaconsDemo extends Component<Props, State> {
     // ])
     // start iBeacon detection
     Beacons.addIBeaconsDetection()
-    .then(() => Beacons.addEddystoneUIDDetection())
-    .then(() => Beacons.addEddystoneURLDetection())
-    .then(() => Beacons.addEddystoneTLMDetection())
-    .then(() => Beacons.addAltBeaconsDetection())
-    .then(() => Beacons.addEstimotesDetection())
-    .catch(error => console.log(`something went wrong during initialization: ${error}`));
-  }
-
-  componentDidMount() {
+      .then(() => Beacons.addEddystoneUIDDetection())
+      .then(() => Beacons.addEddystoneURLDetection())
+      .then(() => Beacons.addEddystoneTLMDetection())
+      .then(() => Beacons.addAltBeaconsDetection())
+      .then(() => Beacons.addEstimotesDetection())
+      .catch(error =>
+        console.log(`something went wrong during initialization: ${error}`),
+      );
     //
     // component state aware here - attach events
     //
@@ -127,30 +133,36 @@ class BeaconsDemo extends Component<Props, State> {
       () => {
         console.log('service connected');
         this.startRangingAndMonitoring();
-      }
+      },
     );
 
     // Ranging: Listen for beacon changes
     this.beaconsDidRangeEvent = Beacons.BeaconsEventEmitter.addListener(
       'beaconsDidRange',
-      (response: {beacons: Array<{distance: number, proximity: string, rssi: string, uuid: string}>, uuid: string, indetifier: string}) => {
+      (response: {
+        beacons: Array<{
+          distance: number,
+          proximity: string,
+          rssi: string,
+          uuid: string,
+        }>,
+        uuid: string,
+        indetifier: string,
+      }) => {
         console.log('BEACONS: ', response);
 
-        response.beacons.forEach(
-          beacon => this.updateBeaconState(
-            RANGING_SECTION_ID,
-            {
-              identifier: response.identifier,
-              uuid: String(beacon.uuid),
-              major: parseInt(beacon.major, 10) >= 0 ? beacon.major : '',
-              minor: parseInt(beacon.minor, 10) >= 0 ? beacon.minor : '',
-              proximity: beacon.proximity ? beacon.proximity : '',
-              rssi: beacon.rssi ? beacon.rssi : '',
-              distance: beacon.distance ? beacon.distance : ''
-            }
-          )
+        response.beacons.forEach(beacon =>
+          this.updateBeaconState(RANGING_SECTION_ID, {
+            identifier: response.identifier,
+            uuid: String(beacon.uuid),
+            major: parseInt(beacon.major, 10) >= 0 ? beacon.major : '',
+            minor: parseInt(beacon.minor, 10) >= 0 ? beacon.minor : '',
+            proximity: beacon.proximity ? beacon.proximity : '',
+            rssi: beacon.rssi ? beacon.rssi : '',
+            distance: beacon.distance ? beacon.distance : '',
+          }),
         );
-      }
+      },
     );
 
     // monitoring:
@@ -158,16 +170,26 @@ class BeaconsDemo extends Component<Props, State> {
       'regionDidEnter',
       ({ identifier, uuid, minor, major }) => {
         console.log('regionDidEnter: ', { identifier, uuid, minor, major });
-        this.updateBeaconState(MONITORING_ENTER_SECTION_ID, { identifier, uuid, minor, major });
-      }
+        this.updateBeaconState(MONITORING_ENTER_SECTION_ID, {
+          identifier,
+          uuid,
+          minor,
+          major,
+        });
+      },
     );
 
     this.beaconsDidLeaveEvent = Beacons.BeaconsEventEmitter.addListener(
       'regionDidExit',
       ({ identifier, uuid, minor, major }) => {
         console.log('regionDidExit: ', { identifier, uuid, minor, major });
-        this.updateBeaconState(MONITORING_LEAVE_SECTION_ID, { identifier, uuid, minor, major });
-      }
+        this.updateBeaconState(MONITORING_LEAVE_SECTION_ID, {
+          identifier,
+          uuid,
+          minor,
+          major,
+        });
+      },
     );
   }
 
@@ -181,7 +203,7 @@ class BeaconsDemo extends Component<Props, State> {
   }
 
   render() {
-    const { beacons } =  this.state;
+    const { beacons } = this.state;
 
     console.log('beacons: ', beacons);
 
@@ -191,26 +213,20 @@ class BeaconsDemo extends Component<Props, State> {
         resizeMode="center"
         source={require('./bluetooth-300-300-opacity-45.png')}
       >
-        <View
-          style={styles.container}
-        >
+        <View style={styles.container}>
           <View style={styles.actionsContainer}>
             <TouchableHighlight
               style={styles.actionButton}
               onPress={this.handlesOnRemoveIbeacon}
             >
-              <Text style={styles.actionText}>
-                remove IBeacon detection
-              </Text>
+              <Text style={styles.actionText}>remove IBeacon detection</Text>
             </TouchableHighlight>
 
             <TouchableHighlight
               style={styles.actionButton}
               onPress={this.handlesOnAddIbeacon}
             >
-              <Text style={styles.actionText}>
-                add IBeacon detection
-              </Text>
+              <Text style={styles.actionText}>add IBeacon detection</Text>
             </TouchableHighlight>
           </View>
 
@@ -228,48 +244,25 @@ class BeaconsDemo extends Component<Props, State> {
       </Image>
     );
   }
+  // #endregion
 
-  sectionListKeyExtractor = (
-    item: DetectedBeacon,
-    index: number
-  ) => {
-    const UUID  = item.uuid ? item.uuid : 'NONE';
-    const ID    = item.identifier ? item.identifier : 'NONE';
+  // #region SectionList related
+  sectionListKeyExtractor = (item: DetectedBeacon, index: number) => {
+    const UUID = item.uuid ? item.uuid : 'NONE';
+    const ID = item.identifier ? item.identifier : 'NONE';
+
     return `${UUID}-${ID}`;
-  }
+  };
 
-  renderSeparator= () => (<View style={{ height: 1, backgroundColor: '#E1E1E1', marginLeft: 80 }} />);
-
-  updateBeaconState = (
-    forSectionId: number = 0, // section identifier
-    {identifier, uuid, minor, major, ...rest} // beacon
-  ) => {
-    const { beacons }    = this.state;
-    const time           = moment().format(TIME_FORMAT);
-    const updatedBeacons = beacons.map(
-                              beacon => {
-                                if (beacon.sectionId === forSectionId) {
-                                  const sameBeacon  = data => !((data.UUID === uuid) && (data.identifier === identifier) && (data.minor === minor) && (data.major === major));
-                                  const updatedData = [].concat(...beacon.data.filter(sameBeacon), { identifier, uuid, minor, major, time, ...rest });
-                                  return { ...beacon, data: updatedData };
-                                }
-                                return beacon;
-                              }
-                            );
-    this.setState({ beacons: updatedBeacons });
-  }
-
-  renderHeader = (
-    { section }
-  ) => (
-    <Text style={styles.headline}>
-      {section.title}
-    </Text>
+  renderHeader = ({ section }) => (
+    <Text style={styles.headline}>{section.title}</Text>
   );
 
-  renderRow = (
-    { item }
-  ) => {
+  renderSeparator = () => (
+    <View style={{ height: 1, backgroundColor: '#E1E1E1', marginLeft: 80 }} />
+  );
+
+  renderRow = ({ item }) => {
     console.log('rowData: ', item);
     return (
       <View style={styles.row}>
@@ -284,19 +277,19 @@ class BeaconsDemo extends Component<Props, State> {
         </View>
         <View style={styles.infoContainer}>
           <Text style={styles.smallText}>
-            indentifier: {item.identifier ? item.identifier  : 'NA'}
+            indentifier: {item.identifier ? item.identifier : 'NA'}
           </Text>
           <Text style={styles.smallText}>
-            UUID: {item.uuid ? item.uuid  : 'NA'}
+            UUID: {item.uuid ? item.uuid : 'NA'}
           </Text>
           <View style={styles.majorMinorContainer}>
             <Text style={styles.smallText}>
               Major: {parseInt(item.major, 10) >= 0 ? item.major : 'NA'}
             </Text>
-            <Text style={[styles.smallText, {marginLeft: 10}]}>
+            <Text style={[styles.smallText, { marginLeft: 10 }]}>
               Minor: {parseInt(item.minor, 10) >= 0 ? item.minor : 'NA'}
             </Text>
-            <Text style={[styles.smallText, {marginLeft: 10}]}>
+            <Text style={[styles.smallText, { marginLeft: 10 }]}>
               RSSI: {item.rssi ? item.rssi : 'NA'}
             </Text>
           </View>
@@ -310,18 +303,50 @@ class BeaconsDemo extends Component<Props, State> {
         </View>
       </View>
     );
-  }
+  };
 
   renderEmpty = () => (
-    <View style={{ height: 40, alignItems: 'center', justifyContent: 'center' }}>
-      <Text>
-        no beacon detected
-      </Text>
+    <View
+      style={{ height: 40, alignItems: 'center', justifyContent: 'center' }}
+    >
+      <Text>no beacon detected</Text>
     </View>
   );
+  // #endregion
+
+  updateBeaconState = (
+    forSectionId: number = 0, // section identifier
+    { identifier, uuid, minor, major, ...rest }, // beacon
+  ) => {
+    const { beacons } = this.state;
+    const time = moment().format(TIME_FORMAT);
+    const updatedBeacons = beacons.map(beacon => {
+      if (beacon.sectionId === forSectionId) {
+        const sameBeacon = data =>
+          !(
+            data.UUID === uuid &&
+            data.identifier === identifier &&
+            data.minor === minor &&
+            data.major === major
+          );
+
+        const updatedData = [].concat(...beacon.data.filter(sameBeacon), {
+          identifier,
+          uuid,
+          minor,
+          major,
+          time,
+          ...rest,
+        });
+        return { ...beacon, data: updatedData };
+      }
+      return beacon;
+    });
+    this.setState({ beacons: updatedBeacons });
+  };
 
   startRangingAndMonitoring = async () => {
-    const { identifier, uuid} = this.state;
+    const { identifier, uuid } = this.state;
     const region = { identifier, uuid }; // minor and major are null here
 
     try {
@@ -332,10 +357,10 @@ class BeaconsDemo extends Component<Props, State> {
     } catch (error) {
       throw error;
     }
-  }
+  };
 
   stopRangingAndMonitoring = async () => {
-    const { identifier, uuid} = this.state;
+    const { identifier, uuid } = this.state;
     const region = { identifier, uuid }; // minor and major are null here
 
     try {
@@ -346,29 +371,43 @@ class BeaconsDemo extends Component<Props, State> {
     } catch (error) {
       throw error;
     }
-  }
+  };
 
   handlesOnAddIbeacon = async () => {
     try {
       await this.stopRangingAndMonitoring();
       await Beacons.addIBeaconsDetection();
       await this.startRangingAndMonitoring();
-      ToastAndroid.showWithGravity('add IBeacon detection', ToastAndroid.SHORT, ToastAndroid.CENTER);
+      ToastAndroid.showWithGravity(
+        'add IBeacon detection',
+        ToastAndroid.SHORT,
+        ToastAndroid.CENTER,
+      );
     } catch (error) {
-      ToastAndroid.show(`Error: add IBeacon detection failed: ${error.message}`, ToastAndroid.SHORT);
+      ToastAndroid.show(
+        `Error: add IBeacon detection failed: ${error.message}`,
+        ToastAndroid.SHORT,
+      );
     }
-  }
+  };
 
   handlesOnRemoveIbeacon = async () => {
     try {
       await this.stopRangingAndMonitoring();
       await Beacons.removeIBeaconsDetection();
       await this.startRangingAndMonitoring();
-      ToastAndroid.showWithGravity('removed IBeacon detection', ToastAndroid.SHORT, ToastAndroid.CENTER);
+      ToastAndroid.showWithGravity(
+        'removed IBeacon detection',
+        ToastAndroid.SHORT,
+        ToastAndroid.CENTER,
+      );
     } catch (error) {
-      ToastAndroid.show(`Error: remove IBeacon detection failed: ${error.message}`, ToastAndroid.SHORT);
+      ToastAndroid.show(
+        `Error: remove IBeacon detection failed: ${error.message}`,
+        ToastAndroid.SHORT,
+      );
     }
-  }
+  };
 }
 
 const styles = StyleSheet.create({
@@ -378,7 +417,7 @@ const styles = StyleSheet.create({
     height: null,
   },
   container: {
-    flex: 1
+    flex: 1,
   },
   btleConnectionStatus: {
     // fontSize: 20,
@@ -386,33 +425,33 @@ const styles = StyleSheet.create({
   },
   headline: {
     fontSize: 20,
-    marginHorizontal: 5
+    marginHorizontal: 5,
   },
   row: {
     flexDirection: 'row',
     padding: 8,
-    paddingBottom: 16
+    paddingBottom: 16,
   },
   iconContainer: {
     flexDirection: 'column',
-    marginRight: 10
+    marginRight: 10,
   },
   infoContainer: {
     flex: 1,
     flexDirection: 'column',
   },
   majorMinorContainer: {
-    flexDirection: 'row'
+    flexDirection: 'row',
   },
   smallText: {
-    fontSize: 11
+    fontSize: 11,
   },
   actionsContainer: {
     marginVertical: 10,
     marginHorizontal: 5,
     flexDirection: 'row',
     alignItems: 'center',
-    justifyContent: 'space-between'
+    justifyContent: 'space-between',
   },
   actionButton: {
     width: 160,
@@ -423,11 +462,8 @@ const styles = StyleSheet.create({
   actionText: {
     alignSelf: 'center',
     fontSize: 11,
-    color: '#F1F1F1'
-  }
+    color: '#F1F1F1',
+  },
 });
 
-AppRegistry.registerComponent(
-  'BeaconsDemo',
-  () => BeaconsDemo
-);
+AppRegistry.registerComponent('BeaconsDemo', () => BeaconsDemo);
