@@ -1,50 +1,59 @@
-/**
-* Sample React Native App
-* https://github.com/facebook/react-native
-* @flow weak
-*/
+// @flow
 
-import React, {
- Component
-}                             from 'react';
-import {
- AppRegistry,
- StyleSheet,
- View,
- Text,
- ListView
-}                             from 'react-native';
-import Beacons                from 'react-native-beacons-manager';
-import BluetoothState         from 'react-native-bluetooth-state';
-import moment                 from 'moment';
+// #region imports
+import React, { Component } from 'react';
+import { AppRegistry, StyleSheet, View, Text, ListView } from 'react-native';
+import Beacons from 'react-native-beacons-manager';
+import BluetoothState from 'react-native-bluetooth-state';
+import moment from 'moment';
+import { hashCode, deepCopyBeaconsLists } from './helpers';
 
-/**
-* uuid of YOUR BEACON (change to yours)
-* @type {String} uuid
-*/
-const UUID        = '7b44b47b-52a1-5381-90c2-f09b6838c5d4';
-const IDENTIFIER  = '123456';
+// #endregion
+
+// #region flow types
+type DetectedBeacon = {
+  identifier: string,
+  uuid?: string,
+  major?: number,
+  minor?: number,
+  proximity?: string,
+  rssi?: string,
+  distance?: number,
+};
+
+type Section = {
+  key: number,
+  data: Array<DetectedBeacon>,
+  title: string,
+  sectionId: string,
+};
+
+type Props = any;
+
+type State = {
+  // region information
+  uuid?: string,
+  identifier: string,
+  // all detected beacons:
+  beacons: Array<Section>,
+};
+// #endregion
+
+// #region constants
+
+// uuid of YOUR BEACON (change to yours)
+const UUID = '7b44b47b-52a1-5381-90c2-f09b6838c5d4';
+const IDENTIFIER = '123456';
 const TIME_FORMAT = 'HH:mm:ss';
 const EMPTY_BEACONS_LISTS = {
-  rangingList:      [],
+  rangingList: [],
   monitorEnterList: [],
-  monitorExitList:  []
+  monitorExitList: [],
 };
 
-const deepCopyBeaconsLists = beaconsLists => {
-  const initial = {};
-  return Object
-          .keys(beaconsLists)
-          .map(key => ({ [key]: [...beaconsLists[key]] }))
-          .reduce(
-            (prev, next) => {
-              return {...prev, ...next};
-            }
-            , initial
-          );
-};
+// #endregion
 
-class BeaconsDemo extends Component {
+class BeaconsDemo extends Component<Props, State> {
   // will be set as list of beacons to update state
   _beaconsLists = null;
 
@@ -59,7 +68,7 @@ class BeaconsDemo extends Component {
 
   state = {
     // region information
-    uuid:       UUID,
+    uuid: UUID,
     identifier: IDENTIFIER,
 
     // check bluetooth state:
@@ -69,11 +78,11 @@ class BeaconsDemo extends Component {
 
     beaconsLists: new ListView.DataSource({
       rowHasChanged: (r1, r2) => r1 !== r2,
-      sectionHeaderHasChanged: (s1, s2) => s1 !== s2
-    }).cloneWithRowsAndSections(EMPTY_BEACONS_LISTS)
+      sectionHeaderHasChanged: (s1, s2) => s1 !== s2,
+    }).cloneWithRowsAndSections(EMPTY_BEACONS_LISTS),
   };
 
-  componentWillMount(){
+  componentDidMount() {
     this._beaconsLists = EMPTY_BEACONS_LISTS;
     const { identifier, uuid } = this.state;
 
@@ -87,67 +96,93 @@ class BeaconsDemo extends Component {
     // (minor and major properties are numbers)
     const region = { identifier, uuid };
     // Monitor for beacons inside the region
-    Beacons
-    .startMonitoringForRegion(region) // or like  < v1.0.7: .startRangingBeaconsInRegion(identifier, uuid)
-    .then(() => console.log('Beacons monitoring started succesfully'))
-    .catch(error => console.log(`Beacons monitoring not started, error: ${error}`));
+    Beacons.startMonitoringForRegion(region) // or like  < v1.0.7: .startRangingBeaconsInRegion(identifier, uuid)
+      .then(() => console.log('Beacons monitoring started succesfully'))
+      .catch(error =>
+        console.log(`Beacons monitoring not started, error: ${error}`),
+      );
 
     // Range for beacons inside the region
-    Beacons
-    .startRangingBeaconsInRegion(region) // or like  < v1.0.7: .startRangingBeaconsInRegion(identifier, uuid)
-    .then(() => console.log('Beacons ranging started succesfully'))
-    .catch(error => console.log(`Beacons ranging not started, error: ${error}`));
+    Beacons.startRangingBeaconsInRegion(region) // or like  < v1.0.7: .startRangingBeaconsInRegion(identifier, uuid)
+      .then(() => console.log('Beacons ranging started succesfully'))
+      .catch(error =>
+        console.log(`Beacons ranging not started, error: ${error}`),
+      );
 
     // update location to be able to monitor:
     Beacons.startUpdatingLocation();
-  }
 
-  componentDidMount() {
     // OPTIONAL: listen to authorization change
     this.authStateDidRangeEvent = Beacons.BeaconsEventEmitter.addListener(
       'authorizationStatusDidChange',
-      (info) => console.log('authorizationStatusDidChange: ', info)
+      info => console.log('authorizationStatusDidChange: ', info),
     );
 
     // Ranging: Listen for beacon changes
     this.beaconsDidRangeEvent = Beacons.BeaconsEventEmitter.addListener(
       'beaconsDidRange',
-      (data) => {
-        this.setState({ message:  'beaconsDidRange event'});
+      data => {
+        this.setState({ message: 'beaconsDidRange event' });
         // console.log('beaconsDidRange, data: ', data);
-        const updatedBeaconsLists = this.updateBeaconList(data.beacons, 'rangingList');
+        const updatedBeaconsLists = this.updateBeaconList(
+          data.beacons,
+          'rangingList',
+        );
         this._beaconsLists = updatedBeaconsLists;
-        this.setState({ beaconsLists: this.state.beaconsLists.cloneWithRowsAndSections(this._beaconsLists)});
-      }
+        this.setState({
+          beaconsLists: this.state.beaconsLists.cloneWithRowsAndSections(
+            this._beaconsLists,
+          ),
+        });
+      },
     );
 
     // monitoring events
     this.regionDidEnterEvent = Beacons.BeaconsEventEmitter.addListener(
       'regionDidEnter',
-      ({uuid, identifier}) => {
-        this.setState({ message:  'regionDidEnter event'});
-        console.log('regionDidEnter, data: ', {uuid, identifier});
+      ({ uuid, identifier }) => {
+        this.setState({ message: 'regionDidEnter event' });
+        console.log('regionDidEnter, data: ', { uuid, identifier });
         const time = moment().format(TIME_FORMAT);
-        const updatedBeaconsLists = this.updateBeaconList({uuid, identifier, time}, 'monitorEnterList');
+        const updatedBeaconsLists = this.updateBeaconList(
+          { uuid, identifier, time },
+          'monitorEnterList',
+        );
         this._beaconsLists = updatedBeaconsLists;
-        this.setState({ beaconsLists: this.state.beaconsLists.cloneWithRowsAndSections(this._beaconsLists)});
-      }
+        this.setState({
+          beaconsLists: this.state.beaconsLists.cloneWithRowsAndSections(
+            this._beaconsLists,
+          ),
+        });
+      },
     );
 
     this.regionDidExitEvent = Beacons.BeaconsEventEmitter.addListener(
       'regionDidExit',
       ({ identifier, uuid, minor, major }) => {
-        this.setState({ message:  'regionDidExit event'});
-        console.log('regionDidExit, data: ', {identifier, uuid, minor, major});
+        this.setState({ message: 'regionDidExit event' });
+        console.log('regionDidExit, data: ', {
+          identifier,
+          uuid,
+          minor,
+          major,
+        });
         const time = moment().format(TIME_FORMAT);
-        const updatedBeaconsLists = this.updateBeaconList({ identifier, uuid, minor, major, time }, 'monitorExitList');
+        const updatedBeaconsLists = this.updateBeaconList(
+          { identifier, uuid, minor, major, time },
+          'monitorExitList',
+        );
         this._beaconsLists = updatedBeaconsLists;
-        this.setState({ beaconsLists: this.state.beaconsLists.cloneWithRowsAndSections(this._beaconsLists)});
-      }
+        this.setState({
+          beaconsLists: this.state.beaconsLists.cloneWithRowsAndSections(
+            this._beaconsLists,
+          ),
+        });
+      },
     );
     // listen bluetooth state change event
-    BluetoothState.subscribe(
-      bluetoothState => this.setState({ bluetoothState: bluetoothState })
+    BluetoothState.subscribe(bluetoothState =>
+      this.setState({ bluetoothState: bluetoothState }),
     );
     BluetoothState.initialize();
   }
@@ -158,16 +193,18 @@ class BeaconsDemo extends Component {
     const region = { identifier, uuid }; // minor and major are null here
 
     // stop monitoring beacons:
-    Beacons
-    .stopMonitoringForRegion(region)
-    .then(() => console.log('Beacons monitoring stopped succesfully'))
-    .catch(error => console.log(`Beacons monitoring not stopped, error: ${error}`));
+    Beacons.stopMonitoringForRegion(region)
+      .then(() => console.log('Beacons monitoring stopped succesfully'))
+      .catch(error =>
+        console.log(`Beacons monitoring not stopped, error: ${error}`),
+      );
 
     // stop ranging beacons:
-    Beacons
-    .stopRangingBeaconsInRegion(region)
-    .then(() => console.log('Beacons ranging stopped succesfully'))
-    .catch(error => console.log(`Beacons ranging not stopped, error: ${error}`));
+    Beacons.stopRangingBeaconsInRegion(region)
+      .then(() => console.log('Beacons ranging stopped succesfully'))
+      .catch(error =>
+        console.log(`Beacons ranging not stopped, error: ${error}`),
+      );
 
     // stop updating locationManager:
     Beacons.stopUpdatingLocation();
@@ -181,42 +218,35 @@ class BeaconsDemo extends Component {
   }
 
   render() {
-    const {
-      bluetoothState,
-      beaconsLists,
-      message
-    } = this.state;
+    const { bluetoothState, beaconsLists, message } = this.state;
 
     return (
       <View style={styles.container}>
         <Text style={styles.btleConnectionStatus}>
-          Bluetooth connection status: { bluetoothState ? bluetoothState  : 'NA' }
+          Bluetooth connection status: {bluetoothState ? bluetoothState : 'NA'}
         </Text>
 
-        <Text>
-          { message }
-        </Text>
+        <Text>{message}</Text>
 
         <View style={styles.justFlex}>
           <ListView
-            dataSource={ beaconsLists }
-            enableEmptySections={ true }
+            dataSource={beaconsLists}
+            enableEmptySections={true}
             renderRow={this.renderBeaconRow}
             renderSectionHeader={this.renderBeaconSectionHeader}
           />
         </View>
-
-       </View>
+      </View>
     );
   }
 
-  renderBeaconRow = (rowData) => (
+  renderBeaconRow = rowData => (
     <View style={styles.row}>
       <Text style={styles.smallText}>
         Identifier: {rowData.identifier ? rowData.identifier : 'NA'}
       </Text>
       <Text style={styles.smallText}>
-        UUID: {rowData.uuid ? rowData.uuid  : 'NA'}
+        UUID: {rowData.uuid ? rowData.uuid : 'NA'}
       </Text>
       <Text style={styles.smallText}>
         Major: {rowData.major ? rowData.major : 'NA'}
@@ -225,7 +255,7 @@ class BeaconsDemo extends Component {
         Minor: {rowData.minor ? rowData.minor : 'NA'}
       </Text>
       <Text style={styles.smallText}>
-        time: { rowData.time ? rowData.time : 'NA'}
+        time: {rowData.time ? rowData.time : 'NA'}
       </Text>
       <Text style={styles.smallText}>
         RSSI: {rowData.rssi ? rowData.rssi : 'NA'}
@@ -237,32 +267,37 @@ class BeaconsDemo extends Component {
         Distance: {rowData.accuracy ? rowData.accuracy.toFixed(2) : 'NA'}m
       </Text>
     </View>
-  )
+  );
 
   renderBeaconSectionHeader = (sectionData, header) => (
-    <Text style={styles.rowSection}>
-       {header}
-     </Text>
-   );
+    <Text style={styles.rowSection}>{header}</Text>
+  );
 
   updateBeaconList = (detectedBeacons = [], listName = '') => {
     // just a deep copy of "this._beaconsLists":
-    const previousLists   = deepCopyBeaconsLists(this._beaconsLists);
-    const listNameIsValid = Object.keys(EMPTY_BEACONS_LISTS).some(header => header === listName);
+    const previousLists = deepCopyBeaconsLists(this._beaconsLists);
+    const listNameIsValid = Object.keys(EMPTY_BEACONS_LISTS).some(
+      header => header === listName,
+    );
     const updateMatchingList = beacon => {
       if (beacon.uuid.length > 0) {
-        const uuid  = beacon.uuid.toUpperCase();
+        const uuid = beacon.uuid.toUpperCase();
         const major = parseInt(beacon.major, 10) ? beacon.major : 0;
         const minor = parseInt(beacon.minor, 10) ? beacon.minor : 0;
 
-        const hasEqualProp = (left, right) => (String(left).toUpperCase() === String(right).toUpperCase());
+        const hasEqualProp = (left, right) =>
+          String(left).toUpperCase() === String(right).toUpperCase();
         const isNotTheSameBeacon = beaconDetail => {
-          return !hasEqualProp(beaconDetail.uuid, uuid)
-              || !hasEqualProp(beaconDetail.major, major)
-              || !hasEqualProp(beaconDetail.minor, minor);
+          return (
+            !hasEqualProp(beaconDetail.uuid, uuid) ||
+            !hasEqualProp(beaconDetail.major, major) ||
+            !hasEqualProp(beaconDetail.minor, minor)
+          );
         };
 
-        const otherBeaconsInSameList = previousLists[listName].filter(isNotTheSameBeacon);
+        const otherBeaconsInSameList = previousLists[listName].filter(
+          isNotTheSameBeacon,
+        );
         previousLists[listName] = [...otherBeaconsInSameList, beacon];
       }
     };
@@ -282,46 +317,43 @@ class BeaconsDemo extends Component {
 
     detectedBeacons.forEach(updateMatchingList);
     return previousLists;
-  }
+  };
 }
 
 const styles = StyleSheet.create({
- container: {
-   flex: 1,
-   paddingTop: 60,
-   margin: 5,
-   backgroundColor: '#F5FCFF',
- },
- justFlex: {
-   flex: 1
- },
- contentContainer: {
-   flex: 1,
-   justifyContent: 'center',
-   alignItems: 'center',
- },
- btleConnectionStatus: {
-   fontSize: 20,
-   paddingTop: 20
- },
- headline: {
-   fontSize: 20,
-   paddingTop: 20,
-   marginBottom: 20
- },
- row: {
-   padding: 8,
-   paddingBottom: 16
- },
-   smallText: {
-   fontSize: 11
- },
- rowSection: {
-   fontWeight: '700'
- }
+  container: {
+    flex: 1,
+    paddingTop: 60,
+    margin: 5,
+    backgroundColor: '#F5FCFF',
+  },
+  justFlex: {
+    flex: 1,
+  },
+  contentContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  btleConnectionStatus: {
+    fontSize: 20,
+    paddingTop: 20,
+  },
+  headline: {
+    fontSize: 20,
+    paddingTop: 20,
+    marginBottom: 20,
+  },
+  row: {
+    padding: 8,
+    paddingBottom: 16,
+  },
+  smallText: {
+    fontSize: 11,
+  },
+  rowSection: {
+    fontWeight: '700',
+  },
 });
 
-AppRegistry.registerComponent(
-  'BeaconsDemo',
-  () => BeaconsDemo
-);
+AppRegistry.registerComponent('BeaconsDemo', () => BeaconsDemo);
