@@ -1,10 +1,10 @@
-// flow weak
+// flow
 
-import { Platform }     from 'react-native';
+import { Platform } from 'react-native';
 
 const RNiBeaconsModule = Platform.select({
-  ios:      () => require('./lib/next/new.module.ios.js'),
-  android:  () => require('./lib/next/new.module.android.js')
+  ios: () => require('./lib/next/new.module.ios.js'),
+  android: () => require('./lib/next/new.module.android.js'),
 })();
 
 export default RNiBeaconsModule;
