@@ -3,11 +3,11 @@
 // #region imports
 const RN = require('react-native');
 
-import type {
-  BeaconRegion,
-  BeaconsManagerANDROID,
-  Parser
-}                           from './module.types';
+import {
+  type BeaconRegion,
+  type BeaconsManagerANDROID,
+  type Parser,
+} from './module.types';
 import {
   PARSER_IBEACON,
   PARSER_ESTIMOTE,
@@ -15,21 +15,20 @@ import {
   PARSER_EDDYSTONE_TLM,
   PARSER_EDDYSTONE_UID,
   PARSER_EDDYSTONE_URL,
-  transmissionSupport
-}                           from './module.types';
+  transmissionSupport,
+} from './module.types';
 // #endregion
 
 // #region instanciation and constants
-const BeaconsManager: BeaconsManagerANDROID = RN.NativeModules.BeaconsAndroidModule;
+const BeaconsManager: BeaconsManagerANDROID =
+  RN.NativeModules.BeaconsAndroidModule;
 const BeaconsEventEmitter = RN.DeviceEventEmitter;
 
-const ARMA_RSSI_FILTER        = BeaconsManager.ARMA_RSSI_FILTER;
+const ARMA_RSSI_FILTER = BeaconsManager.ARMA_RSSI_FILTER;
 const RUNNING_AVG_RSSI_FILTER = BeaconsManager.RUNNING_AVG_RSSI_FILTER;
 // #endregion
 
-function setHardwareEqualityEnforced(
-  flag: boolean
-): void {
+function setHardwareEqualityEnforced(flag: boolean): void {
   BeaconsManager.setHardwareEqualityEnforced(flag);
 }
 
@@ -68,9 +67,9 @@ function removeIBeaconsDetection(): Promise<any> {
 
 // #region altBeacon
 /**
-* set beacon layout for alBeacon
-*
-*/
+ * set beacon layout for alBeacon
+ *
+ */
 function detectAltBeacons(): Promise<any> {
   return new Promise((resolve, reject) => {
     BeaconsManager.addParser(PARSER_ALTBEACON, resolve, reject);
@@ -92,9 +91,9 @@ function removeAltBeaconsDetection(): Promise<any> {
 
 // #region estimote
 /**
-* set beacon layout for estimote
-*
-*/
+ * set beacon layout for estimote
+ *
+ */
 function detectEstimotes(): Promise<any> {
   return new Promise((resolve, reject) => {
     BeaconsManager.addParser(PARSER_ESTIMOTE, resolve, reject);
@@ -116,9 +115,9 @@ function removeEstimotesDetection(): Promise<any> {
 
 // #region eddystone UID
 /**
-* set beacon layout for eddystone UID
-*
-*/
+ * set beacon layout for eddystone UID
+ *
+ */
 function detectEddystoneUID(): Promise<any> {
   return new Promise((resolve, reject) => {
     BeaconsManager.addParser(PARSER_EDDYSTONE_UID, resolve, reject);
@@ -149,9 +148,9 @@ function removeEddystoneUIDDetection(): Promise<any> {
 
 // #region eddystone URL
 /**
-* set beacon layout for eddystone URL
-*
-*/
+ * set beacon layout for eddystone URL
+ *
+ */
 function detectEddystoneURL(): Promise<any> {
   return new Promise((resolve, reject) => {
     BeaconsManager.addParser(PARSER_EDDYSTONE_URL, resolve, reject);
@@ -173,9 +172,9 @@ function removeEddystoneURLDetection(): Promise<any> {
 
 // #region eddystone TLM
 /**
-* set beacon layout for eddystone TLM
-*
-*/
+ * set beacon layout for eddystone TLM
+ *
+ */
 function detectEddystoneTLM(): Promise<any> {
   return new Promise((resolve, reject) => {
     BeaconsManager.addParser(PARSER_EDDYSTONE_TLM, resolve, reject);
@@ -197,28 +196,22 @@ function removeEddystoneTLMDetection(): Promise<any> {
 
 // #region custom beacon (set your parser)
 /**
-* set beacon for custom layout
-*
-*/
-function detectCustomBeaconLayout(
-  parser: number
-): Promise<any> {
+ * set beacon for custom layout
+ *
+ */
+function detectCustomBeaconLayout(parser: number): Promise<any> {
   return new Promise((resolve, reject) => {
     BeaconsManager.addParser(parser, resolve, reject);
   });
 }
 
-function addCustomBeaconLayoutDetection(
-  parser: number
-): Promise<any> {
+function addCustomBeaconLayoutDetection(parser: number): Promise<any> {
   return new Promise((resolve, reject) => {
     BeaconsManager.addParser(parser, resolve, reject);
   });
 }
 
-function removeCustomBeaconLayoutDetection(
-  parser: number
-): Promise<any> {
+function removeCustomBeaconLayoutDetection(parser: number): Promise<any> {
   return new Promise((resolve, reject) => {
     BeaconsManager.removeParser(parser, resolve, reject);
   });
@@ -226,45 +219,32 @@ function removeCustomBeaconLayoutDetection(
 // #endregion
 
 // #region add remove multiple parsers in a row
-function addParsersListToDetection(
-  parsers: Array<Parser>
-) {
+function addParsersListToDetection(parsers: Array<Parser>) {
   return new Promise((resolve, reject) => {
     BeaconsManager.addParsersListToDetection(parsers, resolve, reject);
   });
 }
 
-function removeParsersListToDetection(
-  parsers: Array<Parser>
-) {
+function removeParsersListToDetection(parsers: Array<Parser>) {
   return new Promise((resolve, reject) => {
     BeaconsManager.removeParsersListToDetection(parsers, resolve, reject);
   });
 }
 // #endregion
 
-function setBackgroundScanPeriod(
-  period: number
-): void {
+function setBackgroundScanPeriod(period: number): void {
   BeaconsManager.setBackgroundScanPeriod(period);
 }
 
-function setBackgroundBetweenScanPeriod(
-  period: number
-): void {
+function setBackgroundBetweenScanPeriod(period: number): void {
   BeaconsManager.setBackgroundBetweenScanPeriod(period);
 }
 
-function setForegroundScanPeriod(
-  period: number
-): void {
+function setForegroundScanPeriod(period: number): void {
   BeaconsManager.setForegroundScanPeriod(period);
 }
 
-function setRssiFilter(
-  filterType: number,
-  avgModifier: number
-): void {
+function setRssiFilter(filterType: number, avgModifier: number): void {
   BeaconsManager.setRssiFilter(filterType, avgModifier);
 }
 
@@ -292,11 +272,11 @@ function getMonitoredRegions(): Promise<Array<BeaconRegion>> {
  */
 function checkTransmissionSupported(): Promise<number> {
   return new Promise((resolve, reject) => {
-    BeaconsManager.checkTransmissionSupported(status => resolve(transmissionSupport[status]));
+    BeaconsManager.checkTransmissionSupported(status =>
+      resolve(transmissionSupport[status]),
+    );
   });
 }
-
-
 
 /**
  * start monitoring for a region
@@ -304,22 +284,18 @@ function checkTransmissionSupported(): Promise<number> {
  * @param {Object: BeaconRegion} region region to monitor (identifier + uuid -> major and minor are optional)
  * @returns {Promise<any>} promise resolves to void or error
  */
-function startMonitoringForRegion(
-  region: BeaconRegion
-): Promise<any> {
-  return new Promise(
-    (resolve, reject) => {
-      // NOTE: major and minor are optional values: if user don't assign them we have to send a null value (not undefined):
-      BeaconsManager.startMonitoring(
-        region.identifier,
-        region.uuid,
-        region.minor ? region.minor : -1,
-        region.major ? region.major : -1,
-        resolve,
-        reject
-      );
-    }
-  );
+function startMonitoringForRegion(region: BeaconRegion): Promise<any> {
+  return new Promise((resolve, reject) => {
+    // NOTE: major and minor are optional values: if user don't assign them we have to send a null value (not undefined):
+    BeaconsManager.startMonitoring(
+      region.identifier,
+      region.uuid,
+      region.minor ? region.minor : -1,
+      region.major ? region.major : -1,
+      resolve,
+      reject,
+    );
+  });
 }
 
 /**
@@ -328,21 +304,17 @@ function startMonitoringForRegion(
  * @param {BeaconRegion} region region (see BeaconRegion type)
  * @returns {Promise<any>} promise resolves to void or error
  */
-function stopMonitoringForRegion(
-  region: BeaconRegion
-): Promise<any> {
-  return new Promise(
-    (resolve, reject) => {
-      BeaconsManager.stopMonitoring(
-        region.identifier,
-        region.uuid,
-        region.minor ? region.minor : -1,
-        region.major ? region.major : -1,
-        resolve,
-        reject
-      );
-    }
-  );
+function stopMonitoringForRegion(region: BeaconRegion): Promise<any> {
+  return new Promise((resolve, reject) => {
+    BeaconsManager.stopMonitoring(
+      region.identifier,
+      region.uuid,
+      region.minor ? region.minor : -1,
+      region.major ? region.major : -1,
+      resolve,
+      reject,
+    );
+  });
 }
 
 /**
@@ -354,20 +326,23 @@ function stopMonitoringForRegion(
  */
 function startRangingBeaconsInRegion(
   region: BeaconRegion | string,
-  beaconsUUID?: string
+  beaconsUUID?: string,
 ): Promise<any> {
   if (typeof region === 'object') {
-    return new Promise(
-      (resolve, reject) => {
-        BeaconsManager.startRanging(region.identifier, region.uuid, resolve, reject);
-      }
-    );
+    return new Promise((resolve, reject) => {
+      BeaconsManager.startRanging(
+        // $FlowIgnore
+        region.identifier,
+        // $FlowIgnore
+        region.uuid,
+        resolve,
+        reject,
+      );
+    });
   }
-  return new Promise(
-    (resolve, reject) => {
-      BeaconsManager.startRanging(region, beaconsUUID, resolve, reject);
-    }
-  );
+  return new Promise((resolve, reject) => {
+    BeaconsManager.startRanging(region, beaconsUUID, resolve, reject);
+  });
 }
 
 /**
@@ -379,20 +354,24 @@ function startRangingBeaconsInRegion(
  */
 function stopRangingBeaconsInRegion(
   region: BeaconRegion | string,
-  beaconsUUID?: string
+  beaconsUUID?: string,
 ): Promise<any> {
   if (typeof region === 'object') {
-    return new Promise(
-      (resolve, reject) => {
-        BeaconsManager.stopRanging(region.identifier, region.uuid, resolve, reject);
-      }
-    );
+    return new Promise((resolve, reject) => {
+      BeaconsManager.stopRanging(
+        // $FlowIgnore
+        region.identifier,
+        // $FlowIgnore
+        region.uuid,
+        resolve,
+        reject,
+      );
+    });
   }
-  return new Promise(
-    (resolve, reject) => {
-      BeaconsManager.stopRanging(region, beaconsUUID, resolve, reject);
-    }
-  );
+  return new Promise((resolve, reject) => {
+    // $FlowIgnore
+    BeaconsManager.stopRanging(region, beaconsUUID, resolve, reject);
+  });
 }
 
 module.exports = {
