@@ -2,11 +2,11 @@
 
 const RN = require('react-native');
 
-import type {
-  BeaconRegion,
-  AuthorizationStatus,
-  BeaconsManagerIOS
-}                           from './module.types';
+import {
+  type BeaconRegion,
+  type AuthorizationStatus,
+  type BeaconsManagerIOS,
+} from './module.types';
 
 const BeaconsManager: BeaconsManagerIOS = RN.NativeModules.RNiBeacon;
 const BeaconsEventEmitter = new RN.NativeEventEmitter(BeaconsManager);
@@ -42,9 +42,20 @@ function allowsBackgroundLocationUpdates(allow: boolean = false): void {
  * @returns {() => AuthorizationStatus} instant callback (not async)
  */
 function getAuthorizationStatus(
-  callback: (status: AuthorizationStatus) => any
+  callback: (status: AuthorizationStatus) => any,
 ): any {
   return BeaconsManager.getAuthorizationStatus(callback);
+}
+
+/**
+ * get monitored regions
+ *
+ * @returns {Promise<Array<BeaconRegion>>} promise resolve to an array of monitored regions
+ */
+function getMonitoredRegions(): Promise<Array<BeaconRegion>> {
+  return new Promise((resolve, reject) => {
+    BeaconsManager.getMonitoredRegions(resolve);
+  });
 }
 
 /**
@@ -73,92 +84,76 @@ function shouldDropEmptyRanges(drop: boolean): void {
  * @param {BeaconRegion} region region to monitor (identifier + uuid -> major and minor are optional)
  * @returns {Promise<any>} promise resolves to void or error
  */
-function startMonitoringForRegion(
-  region: BeaconRegion
-): Promise<any> {
-  return new Promise(
-    (resolve, reject) => {
-      try {
-        BeaconsManager.startMonitoringForRegion(region);
-        resolve();
-      } catch (error) {
-        reject(error);
-      }
+function startMonitoringForRegion(region: BeaconRegion): Promise<any> {
+  return new Promise((resolve, reject) => {
+    try {
+      BeaconsManager.startMonitoringForRegion(region);
+      resolve();
+    } catch (error) {
+      reject(error);
     }
-  );
-}
-
-
-/**
-* stop monitoring for a region
-*
-* @param {BeaconRegion} region region (identifier + uuid -> major and minor are optional)
-* @returns {Promise<any>} promise resolves to void or error
-*/
-function stopMonitoringForRegion(
-  region: BeaconRegion
-): Promise<any> {
-  return new Promise(
-    (resolve, reject) => {
-      try {
-        BeaconsManager.stopMonitoringForRegion(region);
-        resolve();
-      } catch (error) {
-        reject(error);
-      }
-    }
-  );
+  });
 }
 
 /**
-* start ranging for a region
-*
-* @param {BeaconRegion} region region to scan (identifier + uuid -> major and minor are optional)
-* @returns {Promise<any>} promise resolves to void or error
-*/
-function startRangingBeaconsInRegion(
-  region: BeaconRegion
-): Promise<any> {
-  return new Promise(
-    (resolve, reject) => {
-      try {
-          BeaconsManager.startRangingBeaconsInRegion(region);
-          resolve();
-        } catch (error) {
-          reject(error);
-        }
+ * stop monitoring for a region
+ *
+ * @param {BeaconRegion} region region (identifier + uuid -> major and minor are optional)
+ * @returns {Promise<any>} promise resolves to void or error
+ */
+function stopMonitoringForRegion(region: BeaconRegion): Promise<any> {
+  return new Promise((resolve, reject) => {
+    try {
+      BeaconsManager.stopMonitoringForRegion(region);
+      resolve();
+    } catch (error) {
+      reject(error);
     }
-  );
+  });
 }
 
 /**
-* stop ranging for a region
-*
-* @param {BeaconRegion} region region (identifier + uuid -> major and minor are optional)
-* @returns {Promise<any>} promise: resolves to void when successful
-*/
-function stopRangingBeaconsInRegion(
-  region: BeaconRegion
-): Promise<any> {
-  return new Promise(
-    (resolve, reject) => {
-      try {
-        BeaconsManager.stopRangingBeaconsInRegion(region);
-        resolve();
-      } catch (error) {
-        reject(error);
-      }
+ * start ranging for a region
+ *
+ * @param {BeaconRegion} region region to scan (identifier + uuid -> major and minor are optional)
+ * @returns {Promise<any>} promise resolves to void or error
+ */
+function startRangingBeaconsInRegion(region: BeaconRegion): Promise<any> {
+  return new Promise((resolve, reject) => {
+    try {
+      BeaconsManager.startRangingBeaconsInRegion(region);
+      resolve();
+    } catch (error) {
+      reject(error);
     }
-  );
+  });
 }
 
-module.exports =  {
+/**
+ * stop ranging for a region
+ *
+ * @param {BeaconRegion} region region (identifier + uuid -> major and minor are optional)
+ * @returns {Promise<any>} promise: resolves to void when successful
+ */
+function stopRangingBeaconsInRegion(region: BeaconRegion): Promise<any> {
+  return new Promise((resolve, reject) => {
+    try {
+      BeaconsManager.stopRangingBeaconsInRegion(region);
+      resolve();
+    } catch (error) {
+      reject(error);
+    }
+  });
+}
+
+module.exports = {
   BeaconsEventEmitter,
 
   requestAlwaysAuthorization,
   requestWhenInUseAuthorization,
   allowsBackgroundLocationUpdates,
   getAuthorizationStatus,
+  getMonitoredRegions,
   startUpdatingLocation,
   stopUpdatingLocation,
   shouldDropEmptyRanges,
@@ -167,5 +162,5 @@ module.exports =  {
   startMonitoringForRegion,
   startRangingBeaconsInRegion,
   stopMonitoringForRegion,
-  stopRangingBeaconsInRegion
+  stopRangingBeaconsInRegion,
 };
