@@ -98,6 +98,21 @@ declare module 'react-native-beacons-manager' {
       region: BeaconRegion
     ): Promise<any>;
 
+    /** IOS ONLY */
+    startRangingBeaconsInRegion(
+      region: BeaconRegion
+    ): Promise<any>;
+
+    /** ANDROID ONLY */
+    startRangingBeaconsInRegion(
+      // We can't simply reuse BeaconRegion as BeaconRegion.uuid is mandatory, whereas the uuid in this method is optional
+      region: {
+        identifier: string,
+        uuid?: string
+      }
+    ): Promise<any>;
+
+    /** ANDROID ONLY */
     startRangingBeaconsInRegion(
       regionId: string,
       beaconsUUID?: string
@@ -107,11 +122,27 @@ declare module 'react-native-beacons-manager' {
       region: BeaconRegion
     ): Promise<any>;
 
+    /** IOS ONLY */
+    stopRangingBeaconsInRegion(
+      region: BeaconRegion
+    ): Promise<any>;
+
+    /** ANDROID ONLY */
     stopRangingBeaconsInRegion(
       regionId: string,
       beaconsUUID?: string
     ): Promise<any>;
+
+    /** ANDROID ONLY */
+    stopRangingBeaconsInRegion(
+      // We can't simply reuse BeaconRegion as BeaconRegion.uuid is mandatory, whereas the uuid in this method is optional
+      region: {
+        identifier: string,
+        uuid?: string
+      }
+    ): Promise<any>;
   }
 
-  export default new Beacons();
+  const beacons: Beacons;
+  export default beacons;
 }
